refactor(ensadmin): move all ai-query-generator code into the nextjs route file
