--- apps/ensadmin/src/components/ai-query-generator/index.ts
@@ -1,1 +1,0 @@
-export * from "./query-generator-client";
